NeighborlyAPI/requirements.txt
flask 
requests 
flask-bootstrap 
flask-restplus 
werkzeug==0.16.1
azure-functions==1.2.1
pymongo==3.10.1
azure-cosmos 
azure-cli-core 
feedgen 
sendgrid